Extract engine preview operation
preview is a bounded media-output operation with scale validation, preview presets, and output probing. This moves it into engine_preview.py while re-exporting preview from mcp_video.engine for compatibility.

Constraint: Existing callers import preview from mcp_video.engine and expect EditResult behavior unchanged.

Rejected: Move storyboard with preview | storyboard has multi-frame/grid output semantics and deserves a separate review slice.

Confidence: high

Scope-risk: narrow

Directive: Keep extracting media-output operations one at a time behind the compatibility facade.

Tested: ruff check --fix mcp_video/engine.py mcp_video/engine_preview.py

Tested: /opt/homebrew/bin/python3 preview re-export smoke

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py -k 'preview' tests/test_server.py -k 'preview' tests/test_cli.py -k 'preview' tests/test_e2e.py -k 'preview' -q --tb=short

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short
